Vision Chatbot/requirements.txt
streamlit
torch
transformers
pillow
